+++ type_convertion/ft_convert_pointer.c
@@ -6,7 +6,7 @@
 /*   By: mbelalou <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/04/17 17:07:43 by mbelalou          #+#    #+#             */
-/*   Updated: 2018/04/21 14:45:34 by mbelalou         ###   ########.fr       */
+/*   Updated: 2018/04/22 02:37:29 by mbelalou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,7 +23,7 @@ static void		put_base(t_format *format, int shift, int size_ret, char *str)
 	ft_put_buf('x', PUT_CHAR);
 	if (format->flags.zero)
 		nbr_char_put = ft_fill_buf(shift, '0');
-	if (format->precision > format->len_temp || shift > 0)
+	if (format->precision > (int)format->len_temp || shift > 0)
 		nbr_char_put += ft_fill_buf(format->precision - format->len_temp, '0');
 	nbr_char_put += ft_put_str(str, format->len_temp);
 	ft_fill_buf(size_ret - nbr_char_put, ' ');
